Refactor questions context
Split dispatch and state. Prevents form re rendering on delete.
